tests: Move make_temp() to ContextManagers
Fixes https://github.com/coala-analyzer/coala/issues/1159
