+++ coalib/misc/ContextManagers.py
@@ -6,6 +6,7 @@ import builtins
 import signal
 import threading
 import platform
+import tempfile
 
 from coalib.misc.MutableValue import MutableValue
 
@@ -146,3 +147,18 @@ def simulate_console_inputs(*inputs):
         yield input_generator
     finally:
         builtins.input = _input
+
+
+@contextmanager
+def make_temp():
+    """
+    Creates a temporary file with a closed stream and deletes it when done.
+
+    :return: A contextmanager retrieving the file path.
+    """
+    temporary = tempfile.mkstemp()
+    os.close(temporary[0])
+    try:
+        yield temporary[1]
+    finally:
+        os.remove(temporary[1])
